Extract `use-audio-snackbar` component
